--- rb_ws/src/buggy/launch/controller.xml
@@ -1,4 +1,0 @@
-<launch>
-    <node pkg="foxglove_bridge" exec="foxglove_bridge" name="foxglove" namespace="SC"/>
-    <node pkg="buggy" exec="controller" name="SC_controller" namespace="SC"/>
-</launch>
