+++ rb_ws/src/buggy/launch/sim_2d_double.xml
@@ -2,24 +2,29 @@
 
     <node pkg="foxglove_bridge" exec="foxglove_bridge" name="foxglove" namespace="SC"/>
     
-    <node pkg="buggy" exec="sim_single" name="SC_sim_single" namespace="SC">
+    <node pkg="buggy" exec="engine.py" name="SC_sim_single" namespace="SC">
         <param name="velocity" value="12"/>
         <param name="pose" value="Hill1_SC"/>
     </node>
-    <node pkg="buggy" exec="buggy_state_converter" name="SC_state_converter" namespace="SC"/>
-    <node pkg="buggy" exec="controller" name="SC_controller" namespace="SC">
+    <node pkg="buggy" exec="buggy_state_converter.py" name="SC_state_converter" namespace="SC"/>
+    <node pkg="buggy" exec="controller_node.py" name="SC_controller" namespace="SC">
         <param name="dist" value="0.0"/>
         <param name="traj_name" value="buggycourse_safe.json"/>
         <param name="controller" value="stanley"/>
     </node>
-    <node pkg="buggy" exec="path_planner" name="SC_path_planner" namespace="SC">
+    <!-- <node pkg="buggy" exec="path_planner" name="SC_path_planner" namespace="SC">
         <param name="traj_name" value="buggycourse_safe.json"/>
-    </node>
+    </node> -->
 
-    <node pkg="buggy" exec="buggy_state_converter" name="NAND_state_converter" namespace="NAND"/>
-    <node pkg="buggy" exec="sim_single" name="NAND_sim_single" namespace="NAND">
-        <param name="velocity" value="12"/>
+    <node pkg="buggy" exec="buggy_state_converter.py" name="NAND_state_converter" namespace="NAND"/>
+    <node pkg="buggy" exec="engine.py" name="NAND_sim_single" namespace="NAND">
+        <param name="velocity" value="10"/>
         <param name="pose" value="Hill1_NAND"/>
+    </node>
+    <node pkg="buggy" exec="controller_node.py" name="NAND_controller" namespace="NAND">
+        <param name="dist" value="0.0"/>
+        <param name="traj_name" value="buggycourse_safe.json"/>
+        <param name="controller" value="stanley"/>
     </node>
 
 </launch>
