--- rb_ws/src/buggy/msg/TrajectoryMsg.msg
@@ -1,4 +1,0 @@
-float64 time
-float64 cur_idx
-float64[] easting
-float64[] northing
