+++ rb_ws/src/buggy/scripts/hello_world.py
@@ -1,5 +1,7 @@
+#!/usr/bin/env python3
+
 def main():
-    print('Hi from buggy. Hello again')
+    print('Hi from buggy.')
 
 
 if __name__ == '__main__':
