+++ rb_ws/src/buggy/scripts/simulator/engine.py
@@ -4,13 +4,13 @@ import sys
 import rclpy
 from rclpy.node import Node
 from geometry_msgs.msg import Pose, Twist, PoseWithCovariance, TwistWithCovariance
-from example_interfaces.msg import Float64
+from std_msgs.msg import Float64
 from sensor_msgs.msg import NavSatFix
 from nav_msgs.msg import Odometry
 import numpy as np
 import utm
-sys.path.append("/rb_ws/src/buggy/buggy")
-from rb_ws.src.buggy.buggy.util.constants import Constants
+sys.path.append("/rb_ws/src/buggy/scripts")
+from util.constants import Constants
 
 class Simulator(Node):
 
@@ -176,6 +176,7 @@ class Simulator(Node):
         if self.tick_count % self.interval == 0:
             self.publish()
         self.tick_count += 1
+        self.get_logger().debug("SIMULATED UTM: ({}, {}), HEADING: {}".format(self.e_utm, self.n_utm, self.heading))
 
 
 def main(args=None):
