--- rb_ws/src/buggy/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/buggy
-[install]
-install_scripts=$base/lib/buggy
